Add 'pass' password manager [1]
Installing 'pass' from brew results in a second gnupg because that is
one of the dependencies. Not a major problem but can be an annoyance
as version differences between gnupg in base and brew do happen. These
difference causes warnings inbetween operations. Having pass in base
lowers my extra package count and prevents doubling of gnupg.

[1]: https://www.passwordstore.org/